Timerも実装

+++ include/sdl/SDLTtfFont.h
@@ -14,8 +14,8 @@ class TtfFont
 {
 public:
 	UNCOPYABLE(TtfFont);
-	TtfFont(const char *file, int ptsize)
-		: font_(::TTF_OpenFont(file, static_cast<float>(ptsize)))
+	TtfFont(const char *file, float ptsize)
+		: font_(::TTF_OpenFont(file, ptsize))
 		, fontSize_(ptsize)
 	{
 	}
@@ -25,7 +25,7 @@ public:
 		}
 	}
 
-	int getFontSize() const { return fontSize_; }
+	float getFontSize() const { return fontSize_; }
 	TTF_Font *get() const { return font_; }
 	bool is_font() const { return font_; }
 
@@ -36,7 +36,7 @@ public:
 
 private:
 	TTF_Font * const font_;
-	const int fontSize_;
+	const float fontSize_;
 };
 
 } // namespace SDL_
